Don't cache Halide_ASAN_ENABLED (#6969)
check_cxx_symbol_exists saves its output in the cache and does not run
if its destination variable is defined. This is OK when used to test
something that necessitates a totally fresh configuration, like any
property of the target architecture, which would require changing the
toolchain file. However, the result here can change if someone just
modifies CMAKE_CXX_FLAGS, so it can get out of sync in some cases.