Add support for generating LLVM vector predication intrinsics. (#7111)
Add support for generating llvm.vp.* intrinsics. This is particularly
useful for RISC V, but it may be a simpler, better optimized path, for
Halide vector operations in general.

Add support for a maximum vector size that might be larger than the
native vector size. RISC V vector LMUL support is an example of an
architecture supporting this.