Disable unreachable-code clang-tidy warnings (#7182)
Some configurations of clang-tidy will (correctly) complain that the code inside the `if` clauses here will never be executed, since it ends up as something like `if (strcmp("foo", "foo"))`... but for testing purposes, we want to keep it, for obvious reasons. It's hard to construct a string-compare here as constexpr, so I'm just going to NOLINT it.

Also changed the `count_buffers()` check to a static_assert for simplicity.