Modify runtime calls to always return a valid halide_error_code_t value (#7404)
* Modify runtime calls to always return a valid halide_error_code_t value

Currently, the return values from our runtime code are a mishmash -- there's lots of code that returns any random nonzero value to indicate an error. This isn't wrong per se, but it's not clean, and it's desirable that the return values are predictable. This PR doesn't change the call signature of any (public) Halide Runtime functions, but modifies the internal logic so that all return values are valid values of `enum halide_error_code_t`.

Generally, there should be minimal change to the code otherwise, although I did leave in a few drive-by changes that I couldn't resist (e.g., better error-checking when dynamically loading symbols).

My long-term goal here is to eventually propose changing the signature of runtime functions that return errors to actually return `enum halide_error_code_t`; as you might imagine, making that transition might be controversial for a number of reasons. This PR is intended to be a way to make such a future transition easier to reason about, while arguably improving the code quality of the runtime slightly.

* tidy

* Update opencl.cpp

* trigger buildbots

* trigger buildbots

* Fix merge mistake

* Update gpu_context_common.h

* Update cuda.cpp

* Werror

* status_

* if-with-initializer format

* Update cuda.cpp

* Update opencl.cpp

* Update cuda.cpp

* Update device_interface.h

* Update hexagon_cache_allocator.cpp

* Update printer.h

* remove prefixes

* "device field is already non-zero"

* Update opencl.cpp

* Update cuda.cpp

* Fix error spacing

* trigger buildbots