Don't inject undef() in the simplifier (#7791)
We shouldn't be using undef() in the simplifier. This replaces a load
with a constant false predicate with a zero instead.

I also added a guard around some dubious logic about out of bounds
loads. out of bounds loads may be reachable if they have a false
predicate, so I changed this simplification to only trigger if the load
is unpredicated.