Handle nested vectorization in store predicates (#7864)
Fix #7851

In one place in PartitionLoops and in another place in the simplifier we
were neglecting to consider nested vectorization.

I added the fuzzer output as a new test, because I have no idea how I'd
generate this error with human-readable code. It stems from an
interaction of several tail strategies.