Improve code size and compile time for local laplacian app (#7927)
Improve code size and compile time for local laplacian and interpolate apps

This reduces compile time for the manual local laplacian schedule from
4.9s to 2.2s, and reduces code size from 126k to 82k

Most of the reduction comes from avoiding a pointless boundary condition
in the output Func. A smaller amount comes from avoiding loop
partitioning using RoundUp and Partition::Never. The Partition::Never
calls are responsible for a 3% reduction in code size and compile times
by themselves.

This has basically no effect on runtime. It seems to reduce it very
slightly, but it's in the noise.